Moved JSON-LD parsing code into JSONLDHelper.
